added-move-preview
When cursor hovers over a unit, will preview that unit's movement. Added
FreeCursorMove for eventual switching between cursor move methods. Note:
Need to keep the update for cursor move functions as fixed update to
make sure that the cursor makes it to the appropriate space by the time
triggers in spaces are checked.